@@ -12,6 +12,22 @@ import Combine
  
     underlying task will receive task cancellation signal if the subscription is cancelled
  
+    
+    **There is an issue when using with PassthroughSubject or CurrentValueSubject**
+ 
+    - Since Swift does not support running Task inline way (run in sync until suspension point), Subject's value can lost.
+    - Use the workaround like below to prevent this kind of issue
+ 
+```
+    import Combine
+
+    let subject = PassthroughSubject<Int,Never>()
+    subject.flatMap(maxPublishers: .max(1)) { value in
+        Just(value).mapTask{ \**do async job**\ }
+    }
+        
+ ```
+ 
  */
 public struct MapTask<Upstream:Publisher, Output:Sendable>: Publisher where Upstream.Output:Sendable {
 
@@ -28,9 +44,9 @@ public struct MapTask<Upstream:Publisher, Output:Sendable>: Publisher where Upst
         }
     }
     
-    public init(upstream: Upstream, transform: @escaping @Sendable (Upstream.Output) async -> Result<Output,Failure>) {
+    public init(upstream: Upstream, handler: @escaping @Sendable (Upstream.Output) async -> Result<Output,Failure>) {
         self.upstream = upstream
-        self.transform = transform
+        self.transform = handler
     }
 
     public func receive<S>(subscriber: S) where S : Subscriber, Upstream.Failure == S.Failure, Output == S.Input {
@@ -48,6 +64,125 @@ extension MapTask: Sendable where Upstream: Sendable {}
 
 extension MapTask {
     
+    
+    internal struct SubscriptionState<S:Subscriber> where S.Input == Output, S.Failure == Failure {
+        
+        typealias AsyncSequenceSource = AsyncMapSequence<AsyncStream<Result<Upstream.Output,Failure>>,Result<Output,Failure>>
+        private let lock: some UnfairStateLock<S?> = createCheckedStateLock(checkedState: nil)
+        private let demandBuffer:DemandAsyncBuffer
+        
+        init(demandBuffer: DemandAsyncBuffer, subscriber:S) {
+            self.demandBuffer = demandBuffer
+            self.lock.withLockUnchecked{ $0 = subscriber }
+        }
+        
+        func receive(_ input: S.Input) -> Subscribers.Demand? {
+            lock.withLockUnchecked{
+                $0
+            }?.receive(input)
+        }
+        
+        func receive(completion: Subscribers.Completion<Failure>) {
+            lock.withLockUnchecked{
+                let old = $0
+                $0 = nil
+                return old
+            }?.receive(completion: completion)
+        }
+        
+        func dropSubscriber() {
+            let _ = lock.withLockUnchecked{
+                let old = $0
+                $0 = nil
+                return old
+            }
+        }
+        
+        func makeSource(
+            upstream:Upstream,
+            transform: @escaping @Sendable (Upstream.Output) async -> Result<Output,Failure>
+        ) async -> (AsyncSequenceSource, Subscription)? {
+            let subscriptionLock = createCheckedStateLock(checkedState: SubscriptionContinuation.waiting)
+            let (stream, continuation) = AsyncStream<Result<Upstream.Output,Failure>>.makeStream()
+            continuation.onTermination = { _ in
+                demandBuffer.close()
+            }
+            upstream
+                .subscribe(
+                    AnySubscriber(
+                        receiveSubscription: subscriptionLock.received,
+                        receiveValue: {
+                            continuation.yield(.success($0))
+                            return .none
+                        },
+                        receiveCompletion: {
+                            switch $0 {
+                            case .finished:
+                                break
+                            case .failure(let error):
+                                continuation.yield(.failure(error))
+                            }
+                            continuation.finish()
+                        }
+                    )
+                )
+            guard let subscription = await subscriptionLock.consumeSubscription()
+            else {
+                continuation.finish()
+                return nil
+            }
+            let source = stream.map{
+                switch $0 {
+                case .success(let success):
+                    return await transform(success)
+                case .failure(let failure):
+                    return .failure(failure)
+                }
+            }
+            return (source,subscription)
+        }
+        
+        func runTask(
+            _ source:AsyncSequenceSource,
+            _ subscription: any Subscription
+        ) async {
+            await withTaskCancellationHandler {
+                var iterator = source.makeAsyncIterator()
+                for await var pending in demandBuffer {
+                    if pending == .none {
+                        subscription.request(.none)
+                        continue
+                    }
+                    while pending > .none {
+                        pending -= 1
+                        subscription.request(.max(1))
+                        guard let result = await iterator.next() else {
+                            return
+                        }
+                        switch result {
+                        case .success(let value):
+                            if let nextDemand = receive(value) {
+                                pending += nextDemand
+                            } else {
+                                return
+                            }
+                        case .failure(let failure):
+                            receive(completion: .failure(failure))
+                            return
+                        }
+                    }
+                }
+            } onCancel: {
+                subscription.cancel()
+                dropSubscriber()
+            }
+            receive(completion: .finished)
+            demandBuffer.close()
+        }
+        
+        
+    }
+    
     private final class Inner<S:Subscriber>:Subscription, CustomStringConvertible, CustomPlaygroundDisplayConvertible where S.Input == Output, S.Failure == Failure {
         
         var description: String { "MapTask" }
@@ -64,56 +199,12 @@ extension MapTask {
             transform: @escaping @Sendable (Upstream.Output) async -> Result<Output,Failure>
         ) {
             let buffer = DemandAsyncBuffer()
-            let lock = createUncheckedStateLock(uncheckedState: S?.some(subscriber))
+            let state = SubscriptionState(demandBuffer: buffer, subscriber: subscriber)
             demander = buffer
             task = Task {
-                let subscriptionLock = createCheckedStateLock(checkedState: SubscriptionContinuation.waiting)
-                let stream = Self.makeStream(
-                    upstream: upstream,
-                    receiveSubscription: subscriptionLock.received,
-                    onCancel: { buffer.close() },
-                    transform: transform
-                )
-                guard let subscription = await subscriptionLock.consumeSubscription()
+                guard let (stream, subscription) = await state.makeSource(upstream: upstream, transform: transform)
                 else { return }
-                await withTaskCancellationHandler {
-                    var iterator = stream.makeAsyncIterator()
-                    for await demand in buffer {
-                        var pending = demand
-                        while pending > .none {
-                            pending -= 1
-                            subscription.request(.max(1))
-                            guard let result = await iterator.next() else {
-                                return
-                            }
-                            switch result {
-                            case .success(let value):
-                                if let currentSubscriber = lock.withLockUnchecked({ $0 }) {
-                                    pending += currentSubscriber.receive(value)
-                                } else {
-                                    return
-                                }
-                            case .failure(let failure):
-                                lock.withLockUnchecked{
-                                    let oldValue = $0
-                                    $0 = nil
-                                    return oldValue
-                                }?.receive(completion: .failure(failure))
-                                return
-                            }
-                        }
-                    }
-
-                } onCancel: {
-                    subscription.cancel()
-                    lock.withLock{ $0 = nil }
-                }
-                lock.withLockUnchecked{
-                    let oldValue = $0
-                    $0 = nil
-                    return oldValue
-                }?.receive(completion: .finished)
-                buffer.close()
+                await state.runTask(stream, subscription)
             }
         }
         
@@ -127,46 +218,7 @@ extension MapTask {
         
         deinit { task.cancel() }
         
-        static func makeStream(
-            upstream:Upstream,
-            receiveSubscription: @escaping (any Subscription) -> Void,
-            onCancel: @escaping @Sendable () -> Void,
-            transform:@escaping @Sendable (Upstream.Output) async -> Result<Output,Failure>
-        ) -> AsyncMapSequence<AsyncStream<Result<Upstream.Output,Failure>>,Result<Output,Failure>> {
-            let stream = AsyncStream<Result<Upstream.Output,Failure>>{ continuation in
-                upstream.subscribe(
-                    AnySubscriber(
-                        receiveSubscription: receiveSubscription,
-                        receiveValue: {
-                            continuation.yield(.success($0))
-                            return .none
-                        },
-                        receiveCompletion: {
-                            switch $0 {
-                            case .finished:
-                                break
-                            case .failure(let error):
-                                continuation.yield(.failure(error))
-                            }
-                            continuation.finish()
-                        }
-                    )
-                )
-                continuation.onTermination = {
-                    if case .cancelled = $0 {
-                        onCancel()
-                    }
-                }
-            }.map{
-                switch $0 {
-                case .success(let success):
-                    return await transform(success)
-                case .failure(let failure):
-                    return .failure(failure)
-                }
-            }
-            return stream
-        }
+
 
     }
     
