@@ -110,88 +110,7 @@ public final class RunLoopScheduler: Scheduler, @unchecked Sendable, Hashable {
         self.config = config
     }
     
-    @usableFromInline
-    internal init(in mode:RunLoop.Mode, config:Configuration) {
-        let commonModes = CFRunLoopCopyAllModes(CFRunLoopGetCurrent()) as! [String]
-        if !commonModes.contains(mode.rawValue) {
-            CFRunLoopAddCommonMode(CFRunLoopGetCurrent(), .init(mode.rawValue as CFString))
-        }
-                
-        let targetRunLoopRef = UnsafeMutablePointer<CFRunLoop?>.allocate(capacity: 1)
-        defer { targetRunLoopRef.deallocate() }
-        targetRunLoopRef.initialize(to: nil)
-        defer { targetRunLoopRef.deinitialize(count: 1) }
-        
-        let emptySource:CFRunLoopSource = {
-            var nullContext = CFRunLoopSourceContext()
-            nullContext.cancel = { _, runLoop, _ in
-                guard let runLoop else { return }
-                CFRunLoopStop(runLoop)
-            }
-            nullContext.copyDescription = { _ in
-                    .passRetained("RunLoopScheduler Default Source" as CFString)
-            }
-            nullContext.version = 0
-            return CFRunLoopSourceCreate(nil, 0, &nullContext)
-        }()
-        
-        let callerRunLoop = RunLoop.current.getCFRunLoop()
-        let callBackSource:CFRunLoopSource = {
-            var context = CFRunLoopSourceContext()
-            context.version = 0
-            context.perform = { _ in
-                
-            }
-            context.copyDescription = { _ in
-                    .passRetained("RunLoopScheduler.InitializerCallBackSource" as CFString)
-            }
-            return CFRunLoopSourceCreate(nil, 0, &context)
-        }()
-        let interruptionObserver = CFRunLoopObserverCreateWithHandler(nil, CFRunLoopActivity.beforeWaiting.union([.beforeTimers, .beforeSources]).rawValue, true, 0) { _, _ in
-            if targetRunLoopRef.pointee != nil {
-                CFRunLoopStop(CFRunLoopGetCurrent())
-            }
-        }.unsafelyUnwrapped
-        CFRunLoopAddSource(CFRunLoopGetCurrent(), callBackSource, .init(mode.rawValue as CFString))
-        CFRunLoopAddObserver(CFRunLoopGetCurrent(), interruptionObserver, .commonModes)
-        let runner = RunLoopRunner(emptySource) {
-            let childLoop = $0
-            
-            CFRunLoopPerformBlock(callerRunLoop, CFRunLoopMode.commonModes.rawValue) {
-                targetRunLoopRef.pointee = childLoop
-                
-                DispatchQueue.global().async {
-                    CFRunLoopSourceSignal(callBackSource)
-                    CFRunLoopWakeUp(callerRunLoop)
-                }
-            }
-            CFRunLoopWakeUp(callerRunLoop)
-        }
-        let workerThread = Thread(target: runner, selector: #selector(runner.main), object: nil)
-        workerThread.qualityOfService = config.qos
-        workerThread.start()
-        while targetRunLoopRef.pointee == nil {
-            RunLoop.current.run(mode: mode, before: .distantFuture)
-            if targetRunLoopRef.pointee != nil {
-                break
-            }
-        }
-        let runLoop = targetRunLoopRef.pointee.unsafelyUnwrapped
-        CFRunLoopObserverInvalidate(interruptionObserver)
-        CFRunLoopSourceInvalidate(callBackSource)
-        self.cfRunLoop = runLoop
-        self.source = emptySource
-        self.config = config
-        
-    }
-    
-    internal static func create(with config:Configuration = .init()) -> RunLoopScheduler {
-        guard let mode = RunLoop.current.currentMode else {
-            return RunLoopScheduler(sync: (), config: config)
-        }
-        return RunLoopScheduler(in: mode, config: config)
-    }
-    
+
     @inlinable
     nonisolated
     public func schedule(
