@@ -38,6 +38,67 @@ public struct AsyncSequencePublisher<Base:AsyncSequence>: Publisher {
 extension AsyncSequencePublisher: Sendable where Base: Sendable, Base.Element: Sendable {}
 
 extension AsyncSequencePublisher {
+    
+    
+    internal struct SubscriptionState<S:Subscriber> where S.Input == Output, S.Failure == Failure {
+        
+        private let lock: some UnfairStateLock<S?> = createUncheckedStateLock(uncheckedState: S?.none)
+        private let demandBuffer:DemandAsyncBuffer
+        
+        init( subscriber: S, demand:DemandAsyncBuffer) {
+            self.demandBuffer = demand
+            lock.withLockUnchecked{ $0 = subscriber }
+        }
+        
+        
+        func receive(_ value: S.Input) -> Subscribers.Demand? {
+            lock.withLockUnchecked{ $0 }?.receive(value)
+        }
+        
+        func receive(completion: Subscribers.Completion<S.Failure>) {
+            lock.withLockUnchecked{
+                let oldValue = $0
+                $0 = nil
+                return oldValue
+            }?.receive(completion: completion)
+        }
+        
+        func dropSubscriber() {
+            // try not to dealloc while holding the lock for safety
+            let _ = lock.withLock{
+                let oldValue = $0
+                $0 = nil
+                return oldValue
+            }
+
+        }
+        
+        func startTask(_ source:consuming Base) async {
+            var iterator = source.makeAsyncIterator()
+            do {
+                for await var pending in demandBuffer {
+                    while pending > .none {
+                        if let value = try await iterator.next() {
+                            pending -= 1
+                            if let newDemand = receive(value) {
+                                pending += newDemand
+                            } else {
+                                return
+                            }
+                        } else {
+                            receive(completion: .finished)
+                            return
+                        }
+                    }
+                }
+                receive(completion: .finished)
+            } catch {
+                receive(completion: .failure(error))
+            }
+        }
+
+        
+    }
 
     private final class Inner<S:Subscriber>: Subscription, CustomStringConvertible, CustomPlaygroundDisplayConvertible, Sendable where S.Input == Output, S.Failure == Failure {
         
@@ -49,13 +110,13 @@ extension AsyncSequencePublisher {
         private let demandBuffer = DemandAsyncBuffer()
         
         internal init(base: Base, subscriber:S) {
-            let lock = createUncheckedStateLock(uncheckedState: Optional<S>.some(subscriber))
             let buffer = demandBuffer
+            let state = SubscriptionState(subscriber: subscriber, demand: buffer)
             self.task = Task {
                 await withTaskCancellationHandler {
-                    await Self.subscribe(lock: lock, demandBuffer: buffer, base: base)
+                    await state.startTask(base)
                 } onCancel: {
-                    lock.withLock{ $0 = nil }
+                    state.dropSubscriber()
                 }
                 buffer.close()
             }
@@ -70,42 +131,7 @@ extension AsyncSequencePublisher {
         }
         
         deinit { task.cancel() }
-        
-        @usableFromInline
-        static func subscribe(
-            lock: some UnfairStateLock<Optional<S>>, demandBuffer: DemandAsyncBuffer, base:Base
-        ) async {
-            var iterator = base.makeAsyncIterator()
-            do {
-            completionLabel:
-                for await demand in demandBuffer {
-                    var pending = demand
-                    while pending > .none {
-                        if let value = try await iterator.next() {
-                            pending -= 1
-                            if let subscriber = lock.withLockUnchecked({ $0 }) {
-                                pending += subscriber.receive(value)
-                            } else {
-                                break completionLabel
-                            }
-                        } else {
-                            break completionLabel
-                        }
-                    }
-                }
-                lock.withLockUnchecked{
-                    let oldValue = $0
-                    $0 = nil
-                    return oldValue
-                }?.receive(completion: .finished)
-            } catch {
-                lock.withLockUnchecked{
-                    let oldValue = $0
-                    $0 = nil
-                    return oldValue
-                }?.receive(completion: .failure(error))
-            }
-        }
     }
+    
     
 }
