@@ -9,49 +9,87 @@ import XCTest
 @testable import Tetra
 
 final class URLSessionDownloadTests: XCTestCase {
+    
+    
+    private static let text = UUID().uuidString
+    private static var webserver:Result<SimpleHTTPServer,Error>? = nil
+    
+    private var url: URL {
+        get throws {
+            let port = try XCTUnwrap(Self.webserver?.get().port?.rawValue)
+            return try XCTUnwrap(URL(string: "http://localhost:\(port)"))
+        }
+    }
+    
+    override class func setUp() {
+        super.setUp()
+        webserver = Result {
+            let server = try SimpleHTTPServer(queue: DispatchQueue(label: "webserver"), port: .any) { error in
+                dump(error)
+                XCTFail(error.localizedDescription)
+            }
+            server.response = text
+            server.start()
+            return server
+        }
+    }
+    
+    override class func tearDown() {
+        super.tearDown()
+        try? webserver?.get().cancel()
+        webserver = nil
+    }
+    
+    override func setUpWithError() throws {
+        try super.setUpWithError()
+        let _ = try Self.webserver?.get()
+    }
+    
 
     func testDefaultDownload() async throws {
-        
-        let (fileURL, response) = try await TetraExtension(base: URLSession.shared).download(
-            from:  URL(string: "https://www.shutterstock.com/image-photo/red-apple-isolated-on-white-260nw-1727544364.jpg")!
-        )
-        let httpResponse = response as! HTTPURLResponse
+        let (fileURL, response) = try await TetraExtension(base: URLSession.shared)
+            .download(from: url)
+        let httpResponse = try XCTUnwrap(response as? HTTPURLResponse)
         addTeardownBlock {
             XCTAssertNoThrow(try FileManager.default.removeItem(at: fileURL))
         }
-        let image = CGImage(jpegDataProviderSource: .init(url: fileURL as CFURL).unsafelyUnwrapped, decode: nil, shouldInterpolate: false, intent: .defaultIntent)
-        XCTAssertNotNil(image)
+        let fileData = try Data(contentsOf: fileURL)
         XCTAssertEqual(httpResponse.statusCode, 200)
+        XCTAssertEqual(fileData, Data(Self.text.utf8))
     }
     
-    func testCancelledDownload() async throws {
+    func testCancelledDownload() async {
         let result = await Task {
             withUnsafeCurrentTask { $0?.cancel() }
-            
-            return try await TetraExtension(base: URLSession.shared).download(
-                from:  URL(string: "https://www.shutterstock.com/image-photo/red-apple-isolated-on-white-260nw-1727544364.jpg")!
-            )
+            return try await TetraExtension(base: URLSession.shared)
+                .download(from: url)
         }.result
         XCTAssertThrowsError(try result.get()) {
-            let urlError = $0 as! URLError
-            XCTAssertEqual(urlError.code, .cancelled)
+            if let urlError = $0 as? URLError {
+                XCTAssertEqual(urlError.code, .cancelled)
+            } else {
+                dump($0)
+                XCTFail($0.localizedDescription)
+            }
         }
     }
     
     func testCancellDuringDownload() async throws {
         let cancelTask2 = Task {
-            try await TetraExtension(base: URLSession.shared).download(
-                from:  URL(string: "https://www.shutterstock.com/image-photo/red-apple-isolated-on-white-260nw-1727544364.jpg")!
-            )
+            try await TetraExtension(base: URLSession.shared)
+                .download(from: url)
         }
-        try await Task.sleep(nanoseconds: 50_000_000)
+        try await Task.sleep(nanoseconds: 20_000_000)
         cancelTask2.cancel()
         let result = await cancelTask2.result
         XCTAssertThrowsError(try result.get()) {
-            let urlError = $0 as! URLError
-            XCTAssertEqual(urlError.code, .cancelled)
+            if let urlError = $0 as? URLError {
+                XCTAssertEqual(urlError.code, .cancelled)
+            } else {
+                dump($0)
+                XCTFail($0.localizedDescription)
+            }
         }
     }
-
 
 }
